Format changes to task lists
Use list-group and form lists of tasks with not details. Start the
process of displaying details in a preview pane. At first, trying to
display to a separate details page, then try to figure out preview pane
later. But having trouble with the new details_page, can't seem to add
its class import to main.py properly. This commit is not functional for
displaying details yet, but does not break the project either. Format
changes to task lists work fine.